fix(Fixed header structure and alignment issues): Fix header-bottom alignment nesting issues on stylesheet and on index page
On the index page, I restructured it so I can see my closing tabs more better as I spent some time finding a missing </div>  in the 'header-bottom' part.  I aslo added some text to introduce my webpage.  On the stylesheet, I grouped the top and bottom parts of the header and added a picture of a cat also in the bottom part.