fix(Restructured the last section of body): Restructure index for section 3 in body, remade the rules in style sheetsheet
So I redid the rules that impacted section 3 in the body, the last section before the header,  to make it so the title, text, and button were alined together. Then I added some background color and rounded the edges of that section. Lastly, I finished the README with a reflection about this project.